Update README.md
Added authorative names of people to research on the theory of the logic I have represented which can be used to help build the logic for anyone.

These ideas are not mine, yours, or theirs. These ideas are shared by all humanity in some form or another. Simple math helps it make sense.